Rework preferences to not be so jank
