+++ ui/src/commonMain/kotlin/cl/emilym/sinatra/ui/localization/TimeFormats.kt
@@ -3,8 +3,9 @@ package cl.emilym.sinatra.ui.localization
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.text.intl.Locale
 import cl.emilym.sinatra.data.models.Time24HSetting
+import cl.emilym.sinatra.data.repository.Preference
 import cl.emilym.sinatra.ui.LanguageConsts
-import cl.emilym.sinatra.ui.widgets.override24HTimeSetting
+import cl.emilym.sinatra.ui.widgets.rememberPreferenceState
 import kotlinx.datetime.LocalDate
 import kotlinx.datetime.LocalDateTime
 import kotlinx.datetime.LocalTime
@@ -26,7 +27,7 @@ import sinatra.ui.generated.resources.time_pm
 
 @Composable
 fun is24HourTimeFormat(): Boolean {
-    val override = override24HTimeSetting()
+    val override by rememberPreferenceState(Preference.Use24HourUnits)
     return when (override) {
         Time24HSetting.OVERRIDE_12 -> false
         Time24HSetting.OVERRIDE_24 -> true
