+++ ui/src/commonMain/kotlin/cl/emilym/sinatra/ui/presentation/screens/preferences/PreferencesScreen.kt
@@ -4,10 +4,8 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.ColumnScope
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
@@ -16,7 +14,6 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -24,24 +21,11 @@ import androidx.compose.ui.unit.dp
 import cafe.adriel.voyager.core.screen.Screen
 import cl.emilym.compose.units.rdp
 import cl.emilym.sinatra.data.models.ContentLink
-import cl.emilym.sinatra.data.models.Time24HSetting
 import cl.emilym.sinatra.data.repository.PreferencesRepository
-import cl.emilym.sinatra.data.repository.StatefulPreferencesUnit
-import cl.emilym.sinatra.data.repository.state
 import cl.emilym.sinatra.nullIfEmpty
 import cl.emilym.sinatra.ui.widgets.ContentLinkWidget
 import cl.emilym.sinatra.ui.widgets.NavigatorBackButton
-import org.koin.compose.getKoin
 import org.koin.compose.koinInject
-
-data class PreferencesCollection(
-    val requiresWheelchair: StatefulPreferencesUnit<Boolean>,
-    val requiresBikes: StatefulPreferencesUnit<Boolean>,
-    val maximumWalkingTime: StatefulPreferencesUnit<Float>,
-    val showAccessibilityIconsNavigation: StatefulPreferencesUnit<Boolean>,
-    val metric: StatefulPreferencesUnit<Boolean>,
-    val time24Hour: StatefulPreferencesUnit<Time24HSetting>
-)
 
 abstract class PreferencesScreen: Screen {
     @get:Composable
@@ -77,16 +61,7 @@ abstract class PreferencesScreen: Screen {
                             .padding(horizontal = 1.rdp),
                         verticalArrangement = Arrangement.spacedBy(2.rdp)
                     ) {
-                        Preferences(
-                            PreferencesCollection(
-                                preferencesRepository.requiresWheelchair.state(scope),
-                                preferencesRepository.requiresBikes.state(scope),
-                                preferencesRepository.maximumWalkingTime.state(scope),
-                                preferencesRepository.showAccessibilityIconsNavigation.state(scope),
-                                preferencesRepository.metric.state(scope),
-                                preferencesRepository.use24Hour.state(scope)
-                            )
-                        )
+                        Preferences()
                     }
                     options().nullIfEmpty()?.let {
                         Column(
@@ -104,9 +79,7 @@ abstract class PreferencesScreen: Screen {
     }
 
     @Composable
-    abstract fun ColumnScope.Preferences(
-        preferencesCollection: PreferencesCollection
-    )
+    abstract fun ColumnScope.Preferences()
 
     @Composable
     open fun options(): List<ContentLink> { return emptyList() }
