+++ ui/src/commonMain/kotlin/cl/emilym/sinatra/ui/presentation/screens/preferences/RoutingPreferencesScreen.kt
@@ -11,6 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import cafe.adriel.voyager.core.screen.ScreenKey
 import cl.emilym.sinatra.FeatureFlag
+import cl.emilym.sinatra.data.repository.Preference
 import cl.emilym.sinatra.lib.FloatRange
 import cl.emilym.sinatra.ui.text
 import cl.emilym.sinatra.ui.widgets.form.HorizontalLockup
@@ -37,7 +38,7 @@ class RoutingPreferencesScreen: PreferencesScreen() {
         get() = stringResource(Res.string.preferences_routing_title)
 
     @Composable
-    override fun ColumnScope.Preferences(preferencesCollection: PreferencesCollection) {
+    override fun ColumnScope.Preferences() {
         val showAccessibilitySettings = !FeatureFlag.GLOBAL_HIDE_TRANSPORT_ACCESSIBILITY.value()
         if (showAccessibilitySettings) {
             HorizontalLockup(
@@ -45,7 +46,7 @@ class RoutingPreferencesScreen: PreferencesScreen() {
                 stringResource(Res.string.preferences_setting_wheelchair_subtitle),
                 Modifier.fillMaxWidth()
             ) {
-                PreferencesCheckbox(preferencesCollection.requiresWheelchair)
+                PreferencesCheckbox(Preference.RequiresWheelchair)
             }
 
             HorizontalLockup(
@@ -53,7 +54,7 @@ class RoutingPreferencesScreen: PreferencesScreen() {
                 stringResource(Res.string.preferences_setting_bikes_subtitle),
                 Modifier.fillMaxWidth()
             ) {
-                PreferencesCheckbox(preferencesCollection.requiresBikes)
+                PreferencesCheckbox(Preference.RequiresBikes)
             }
 
             HorizontalLockup(
@@ -61,7 +62,7 @@ class RoutingPreferencesScreen: PreferencesScreen() {
                 stringResource(Res.string.preferences_setting_show_accessibility_icons_navigation_subtitle),
                 Modifier.fillMaxWidth()
             ) {
-                PreferencesCheckbox(preferencesCollection.showAccessibilityIconsNavigation)
+                PreferencesCheckbox(Preference.ShowAccessibilityIconsNavigation)
             }
         }
 
@@ -71,7 +72,7 @@ class RoutingPreferencesScreen: PreferencesScreen() {
             Modifier.fillMaxWidth()
         ) {
             PreferencesFloatSlider(
-                preferencesCollection.maximumWalkingTime,
+                Preference.MaximumWalkingTime,
                 FloatRange(10f, 60f),
                 Modifier.fillMaxWidth(1f)
             ) {
