+++ ui/src/commonMain/kotlin/cl/emilym/sinatra/ui/widgets/form/PreferencesCheckbox.kt
@@ -1,21 +1,19 @@
 package cl.emilym.sinatra.ui.widgets.form
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Modifier
-import cl.emilym.sinatra.data.repository.StatefulPreferencesUnit
-import cl.emilym.sinatra.ui.widgets.collectAsStateWithLifecycle
-import kotlinx.coroutines.launch
+import cl.emilym.sinatra.data.repository.Preference
+import cl.emilym.sinatra.ui.widgets.rememberPreferenceState
 
 @Composable
 fun PreferencesCheckbox(
-    unit: StatefulPreferencesUnit<Boolean>,
+    preference: Preference<Boolean>,
     modifier: Modifier = Modifier
 ) {
-    val scope = rememberCoroutineScope()
+    var value by rememberPreferenceState(preference)
     SinatraCheckbox(
-        unit.flow.collectAsStateWithLifecycle().value,
-        { scope.launch { unit.save(it) } },
+        value,
+        { value = it },
         modifier
     )
 }
